Add support for creating a bridge device for use by VMs for Ubuntu
This change is Ubuntu specific because it uses [netplan](https://netplan.io/).
It can be made generic for Linux through the use of the [ip command](https://wiki.archlinux.org/title/Network_bridge#With_iproute2),
but this configuration will not be persistent.

The `create-host-bridge` script is also not currently idempotent, and it may
be dangerous to run on Linux boxes with more advanced networking setups besides the default.

The main purpose behind this PR is to allow setting up a single Helios VM that can be used
to run falcon and spin up other helios machines for testing.